format contest

+++ contracts/src/mocks/MockCompound.sol
@@ -8,7 +8,7 @@ import "./MockUSDC.sol";
  * @title MockCompound
  * @dev This contract mocks Compound V3 Comet for testing purposes.
  * It implements the exact interface that our DepositManager expects.
- * 
+ *
  * Real Compound V3 Comet functions we need to mock:
  * - supply(address asset, uint256 amount)
  * - withdraw(address asset, uint256 amount)
@@ -21,38 +21,38 @@ import "./MockUSDC.sol";
  */
 contract MockCompound {
     IERC20 public immutable underlying;
-    
+
     // Balance tracking (like real Compound V3)
     mapping(address => uint256) public balances;
     uint256 public totalSupplyAmount;
-    
+
     // Pause flags (like real Compound V3)
     bool public _isSupplyPaused;
     bool public _isWithdrawPaused;
     bool public _supplyShouldFail;
-    
+
     // Mock rates for testing (can be updated via admin functions)
     uint256 public mockUtilization; // 18 decimals (e.g., 0.5 * 10^18 = 50%)
-    uint256 public mockSupplyRate;  // Per second rate with 18 decimals
-    
+    uint256 public mockSupplyRate; // Per second rate with 18 decimals
+
     // Events (like real Compound V3)
-    event Supply(address indexed from, address indexed dst, uint amount);
-    event Withdraw(address indexed from, address indexed dst, uint amount);
+    event Supply(address indexed from, address indexed dst, uint256 amount);
+    event Withdraw(address indexed from, address indexed dst, uint256 amount);
 
     constructor(address _underlying) {
         underlying = IERC20(_underlying);
-        
+
         // Set default mock values
         // Default: 91.1% utilization (0.911 * 10^18)
         mockUtilization = 911000000000000000; // 91.1%
-        
+
         // Default: ~7% APR supply rate
         // APR of 7% = 0.07 per year
         // Rate per second = 0.07 / 31536000 seconds = 0.00000000221961970319...
         // In wei (18 decimals) = 2219619703 wei/second
         mockSupplyRate = 2219619703;
     }
-    
+
     /**
      * @dev Supply underlying asset to Compound V3
      * @param asset The underlying asset address
@@ -63,14 +63,14 @@ contract MockCompound {
         require(asset == address(underlying), "Invalid asset");
         require(amount > 0, "Amount must be greater than 0");
         require(!_supplyShouldFail, "Supply failed");
-        
+
         // Transfer underlying from user to this contract
         underlying.transferFrom(msg.sender, address(this), amount);
-        
+
         // Update balances (like real Compound V3)
         balances[msg.sender] += amount;
         totalSupplyAmount += amount;
-        
+
         emit Supply(msg.sender, msg.sender, amount);
     }
 
@@ -84,14 +84,14 @@ contract MockCompound {
         require(asset == address(underlying), "Invalid asset");
         require(amount > 0, "Amount must be greater than 0");
         require(balances[msg.sender] >= amount, "Insufficient balance");
-        
+
         // Update balances (like real Compound V3)
         balances[msg.sender] -= amount;
         totalSupplyAmount -= amount;
-        
+
         // Transfer underlying to user
         underlying.transfer(msg.sender, amount);
-        
+
         emit Withdraw(msg.sender, msg.sender, amount);
     }
 
@@ -182,7 +182,7 @@ contract MockCompound {
         // Mint the underlying tokens to this contract first
         // This simulates the yield generation in real Compound V3
         MockUSDC(address(underlying)).mint(address(this), amount);
-        
+
         // Then add to the account balance
         balances[to] += amount;
         totalSupplyAmount += amount;
@@ -203,4 +203,4 @@ contract MockCompound {
     function setMockSupplyRate(uint256 supplyRate) external {
         mockSupplyRate = supplyRate;
     }
-} 
+}
