+++ contracts/test/ContestFactory.t.sol
@@ -9,19 +9,19 @@ import "../src/PlatformToken.sol";
 contract ContestFactoryTest is Test {
     ContestFactory public factory;
     PlatformToken public platformToken;
-    
+
     address public oracle = address(1);
     address public user1 = address(2);
-    
+
     function setUp() public {
         factory = new ContestFactory();
         platformToken = new PlatformToken("Cut Platform Token", "CUT");
     }
-    
+
     function testCreateContest() public {
         uint256 depositAmount = 100e18; // 100 CUT tokens
         uint256 expiry = block.timestamp + 7 days;
-        
+
         address contest = factory.createContest(
             address(platformToken),
             oracle,
@@ -31,23 +31,23 @@ contract ContestFactoryTest is Test {
             1000e18, // liquidity parameter
             500, // demand sensitivity 5%
             750, // prize share 7.5%
-            750  // user share 7.5%
+            750 // user share 7.5%
         );
-        
+
         assertEq(factory.contestHost(contest), address(this));
-        
+
         address[] memory contests = factory.getContests();
         assertEq(contests.length, 1);
         assertEq(contests[0], contest);
         assertEq(factory.getContestCount(), 1);
     }
-    
+
     function testMultipleContests() public {
         uint256 depositAmount = 50e18;
         uint256 expiry = block.timestamp + 7 days;
-        
+
         // Create 3 contests
-        for (uint i = 0; i < 3; i++) {
+        for (uint256 i = 0; i < 3; i++) {
             factory.createContest(
                 address(platformToken),
                 oracle,
@@ -57,14 +57,13 @@ contract ContestFactoryTest is Test {
                 1000e18,
                 500,
                 750, // prize share
-                750  // user share
+                750 // user share
             );
         }
-        
+
         assertEq(factory.getContestCount(), 3);
-        
+
         address[] memory contests = factory.getContests();
         assertEq(contests.length, 3);
     }
 }
-
